[android][ios] Upgrade react-native-pager-view to 6.1.2
