+++ apps/bare-expo/package.json
@@ -83,7 +83,7 @@
     "react-dom": "18.2.0",
     "react-native": "0.71.0",
     "react-native-gesture-handler": "~2.9.0",
-    "react-native-pager-view": "6.0.1",
+    "react-native-pager-view": "6.1.2",
     "react-native-reanimated": "~2.14.0",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.19.0",
