+++ apps/native-component-list/package.json
@@ -144,7 +144,7 @@
     "react-native-dropdown-picker": "^5.3.0",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-maps": "1.3.2",
-    "react-native-pager-view": "6.0.1",
+    "react-native-pager-view": "6.1.2",
     "react-native-paper": "^4.0.1",
     "react-native-reanimated": "~2.14.0",
     "react-native-safe-area-context": "4.5.0",
