+++ packages/expo/bundledNativeModules.json
@@ -92,7 +92,7 @@
   "react-native-gesture-handler": "~2.9.0",
   "react-native-get-random-values": "~1.8.0",
   "react-native-maps": "1.3.2",
-  "react-native-pager-view": "6.0.1",
+  "react-native-pager-view": "6.1.2",
   "react-native-reanimated": "~2.14.0",
   "react-native-screens": "~3.19.0",
   "react-native-safe-area-context": "4.5.0",
